Factorizing shared constants between parsers and serializers
